use directory as file copy path
Add a check to see if the file path provided for file template
destination is a directory (`tim plate` will just copy the file into the
directory). Note that this option currently doesn't work if the
directory does not yet exist (I have no plans to make that work, as it
feels like a poor idea). Updated README.md