switch to urfav/cli
Finally decided to move away from complete self implementation, as I
want this to be a viable tool later on. Using urfav/cli for cli
implementation, and changed the functionality of some of the actions.
File management has changed a bit, and I want to add local plate caching
in the future.

- Changed "Src" based templating to "plate" based templating (for fun?)